[update] untested, but maybe a better way to organize the source
