+++ events/interactionCreate.js
@@ -1,0 +1,19 @@
+module.exports = {
+  name: "interactionCreate",
+  once: false,
+  async execute(interaction) {
+    if (!interaction.isCommand()) return;
+    const command = client.commands.get(interaction.commandName);
+    if (!command) return;
+
+    try {
+      await command.execute(interaction, client);
+    } catch (error) {
+      console.error(error);
+      await interaction.reply({
+        content: "There was an error while executing this command!",
+        ephemeral: true,
+      });
+    }
+  },
+};
